Moved sprindi examples
