+++ spring-core/src/main/java/com/baeldung/constructordi/SpringRunner.java
@@ -1,9 +1,10 @@
-package com.baeldung.spring;
+package com.baeldung.constructordi;
 
-import com.baeldung.spring.domain.Car;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
+
+import com.baeldung.constructordi.domain.Car;
 
 public class SpringRunner {
     public static void main(String[] args) {
