+++ spring-core/src/main/resources/baeldung.xml
@@ -3,17 +3,17 @@
        xsi:schemaLocation="http://www.springframework.org/schema/beans
     http://www.springframework.org/schema/beans/spring-beans-3.0.xsd">
 
-    <bean id="toyota" class="com.baeldung.spring.domain.Car">
+    <bean id="toyota" class="com.baeldung.constructordi.domain.Car">
         <constructor-arg index="0" ref="engine"/>
         <constructor-arg index="1" ref="transmission"/>
     </bean>
 
-    <bean id="engine" class="com.baeldung.spring.domain.Engine">
+    <bean id="engine" class="com.baeldung.constructordi.domain.Engine">
         <constructor-arg index="0" value="v4"/>
         <constructor-arg index="1" value="2"/>
     </bean>
 
-    <bean id="transmission" class="com.baeldung.spring.domain.Transmission">
+    <bean id="transmission" class="com.baeldung.constructordi.domain.Transmission">
         <constructor-arg value="sliding"/>
     </bean>
